Делает рафакторинг: - Разбивает модуль вспомогательных функций - Выделим весь код, что отвечает за отрисовку точек, в отдельную функцию
