+++ src/utils/common.js
@@ -1,0 +1,19 @@
+const getRandomArrayElement = (items) => items[Math.floor(Math.random() * items.length)];
+
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const createIdGenerator = () => {
+  let lastGeneratedId = 0;
+
+  return function () {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+export { getRandomArrayElement, getRandomInteger, createIdGenerator };
